create interfaces for services, changed domain dependency
