+++ jss-data/jss-repository/src/main/resources/db/liquibase-create-table-project.yaml
@@ -7,7 +7,7 @@ databaseChangeLog:
             tableName: project
             columns:
               - column:
-                  name: id
+                  name: project_id
                   type: serial
                   constraints:
                     primaryKey: true
